Convert the framerate status to a floating point value.
During video recording, the framerate can fluctuate as flow control
is applied, so switching to a float gives a bit more precision in
the UI display.

+++ src/pipeline/pipeline.h
@@ -88,7 +88,7 @@ struct pipeline_state {
 
     /* Recording Mode */
     unsigned int    preroll;
-    unsigned int    estrate;
+    double          estrate;
     struct timespec frametime;
 
     /* Pipeline args */
